+++ .gitignore
@@ -44,21 +44,11 @@ coverage/
 
 # 문서 폴더 (Docs)
 docs/
+src/uploads
 uploads/
 
 # 기타
 npm-shrinkwrap.json
 
-src/uploads
-uploads/
 
-# 폴더 전체 무시
-src/routers/*
-src/controllers/*
-src/services/*
-src/validators/*
-src/middlewares/*
 
-# 특정 파일은 예외 처리
-!src/routers/image-router.js
-!src/routers/tag-router.js
