+++ README.md
@@ -1,28 +1,2 @@
 # nb04-How-Do-I-Look-team4
 코드잇 초급 프로젝트
-
-## Instructions
-- 각자 지정된 줄(자기 이름 있는 줄)부터 작성 시작
-- 다른 사람 구역은 절대 건드리지 말 것
-- 줄번호는 편의를 위한 안내일 뿐, 실제 커밋 시 삭제해도 됨
-
----
-
-## Member Notes
-
-1. [13] LSJ: 코멘트 담당입니다. 현재 GIT PR, merge 방법에 대한 과정을 확인하고 있습니다.  
-
-
-2. [16] LHY:  식곤증 끝 당땡김 ㅜㅜㅜ
-
-
-3. [19] KHY:  
-
-
-4. [22] SCJ: 큐레이팅.
-
-
-5. [25] CJH: ...
-
-
-6. [28] CHK:  누가 발로 찼다 아포 서진 선생님 말 잘 듣자!
