--- src/middlewares/errorHandler-middleware.js
@@ -1,8 +1,0 @@
-export default function errorHandler(err, req, res, next) {
-  console.error(err);
-  res.status(err.status || 500).json({
-    success: false,
-    message: err.message || "서버 오류", 
-  });
-}
-
